samples/bot-proactive-messaging/python/coordinate-logger/requirements.txt
# Microsoft Teams SDK packages (v2.0 alpha)
microsoft-teams-ai==2.0.0a8
microsoft-teams-api==2.0.0a8
microsoft-teams-apps==2.0.0a8
microsoft-teams-cards==2.0.0a8
microsoft-teams-common==2.0.0a8

# Environment variables
python-dotenv>=1.0.0
